+++ packages/google_mlkit_text_recognition/lib/src/text_recognizer.dart
@@ -84,12 +84,13 @@ class TextBlock {
   final List<Point<int>> cornerPoints;
 
   /// Constructor to create an instance of [TextBlock].
-  TextBlock(
-      {required this.text,
-      required this.lines,
-      required this.boundingBox,
-      required this.recognizedLanguages,
-      required this.cornerPoints});
+  TextBlock({
+    required this.text,
+    required this.lines,
+    required this.boundingBox,
+    required this.recognizedLanguages,
+    required this.cornerPoints,
+  });
 
   /// Returns an instance of [TextBlock] from a given [json].
   factory TextBlock.fromJson(Map<dynamic, dynamic> json) {
@@ -104,11 +105,12 @@ class TextBlock {
       lines.add(textLine);
     }
     return TextBlock(
-        text: text,
-        lines: lines,
-        boundingBox: rect,
-        recognizedLanguages: recognizedLanguages,
-        cornerPoints: points);
+      text: text,
+      lines: lines,
+      boundingBox: rect,
+      recognizedLanguages: recognizedLanguages,
+      cornerPoints: points,
+    );
   }
 }
 
@@ -129,18 +131,31 @@ class TextLine {
   /// The corner points of the text line in clockwise order starting with the top left point relative to the image in the default coordinate space.
   final List<Point<int>> cornerPoints;
 
+  // The confidence of the recognized line.
+  // Only available in Android, for iOS returns null.
+  final double? confidence;
+
+  // The angle of the rotation of the recognized line.
+  // Only available in Android, for iOS returns null.
+  final double? angle;
+
   /// Constructor to create an instance of [TextLine].
-  TextLine(
-      {required this.text,
-      required this.elements,
-      required this.boundingBox,
-      required this.recognizedLanguages,
-      required this.cornerPoints});
+  TextLine({
+    required this.text,
+    required this.elements,
+    required this.boundingBox,
+    required this.recognizedLanguages,
+    required this.cornerPoints,
+    required this.confidence,
+    required this.angle,
+  });
 
   /// Returns an instance of [TextLine] from a given [json].
   factory TextLine.fromJson(Map<dynamic, dynamic> json) {
     final text = json['text'];
     final rect = RectJson.fromJson(json['rect']);
+    final confidence = json['confidence'];
+    final angle = json['angle'];
     final recognizedLanguages =
         _listToRecognizedLanguages(json['recognizedLanguages']);
     final points = _listToCornerPoints(json['points']);
@@ -150,11 +165,14 @@ class TextLine {
       elements.add(textElement);
     }
     return TextLine(
-        text: text,
-        elements: elements,
-        boundingBox: rect,
-        recognizedLanguages: recognizedLanguages,
-        cornerPoints: points);
+      text: text,
+      elements: elements,
+      boundingBox: rect,
+      recognizedLanguages: recognizedLanguages,
+      cornerPoints: points,
+      confidence: confidence,
+      angle: angle,
+    );
   }
 }
 
@@ -163,24 +181,113 @@ class TextElement {
   /// String representation of the text element that was recognized.
   final String text;
 
+  /// List of text elements that make up the line.
+  // Only available in Android, for iOS returns an empty list.
+  final List<TextSymbol> symbols;
+
   /// Rect that contains the text element.
   final Rect boundingBox;
+
+  /// List of recognized languages in the text element. If no languages were recognized, the list is empty.
+  final List<String> recognizedLanguages;
 
   /// List of corner points of the text element in clockwise order starting with the top left point relative to the image in the default coordinate space.
   final List<Point<int>> cornerPoints;
 
+  // The confidence of the recognized element.
+  // Only available in Android, for iOS returns null.
+  final double? confidence;
+
+  // The angle of the rotation of the recognized element.
+  // Only available in Android, for iOS returns null.
+  final double? angle;
+
   /// Constructor to create an instance of [TextElement].
-  TextElement(
-      {required this.text,
-      required this.boundingBox,
-      required this.cornerPoints});
+  TextElement({
+    required this.text,
+    required this.symbols,
+    required this.boundingBox,
+    required this.recognizedLanguages,
+    required this.cornerPoints,
+    required this.confidence,
+    required this.angle,
+  });
 
   /// Returns an instance of [TextElement] from a given [json].
   factory TextElement.fromJson(Map<dynamic, dynamic> json) {
     final text = json['text'];
     final rect = RectJson.fromJson(json['rect']);
+    final recognizedLanguages =
+        _listToRecognizedLanguages(json['recognizedLanguages']);
     final points = _listToCornerPoints(json['points']);
-    return TextElement(text: text, boundingBox: rect, cornerPoints: points);
+    final confidence = json['confidence'];
+    final angle = json['angle'];
+    final symbols = <TextSymbol>[];
+    for (final symbol in json['symbols']) {
+      final textSymbol = TextSymbol.fromJson(symbol);
+      symbols.add(textSymbol);
+    }
+    return TextElement(
+      text: text,
+      symbols: symbols,
+      boundingBox: rect,
+      recognizedLanguages: recognizedLanguages,
+      cornerPoints: points,
+      confidence: confidence,
+      angle: angle,
+    );
+  }
+}
+
+/// A text symbol recognized in an image. Represents a single symbol in an [TextElement].
+class TextSymbol {
+  /// String representation of the text symbol that was recognized.
+  final String text;
+
+  /// Rect that contains the text symbol.
+  final Rect boundingBox;
+
+  /// List of recognized languages in the text symbol. If no languages were recognized, the list is empty.
+  final List<String> recognizedLanguages;
+
+  /// List of corner points of the text symbol in clockwise order starting with the top left point relative to the image in the default coordinate space.
+  final List<Point<int>> cornerPoints;
+
+  // The confidence of the recognized symbol.
+  // Only available in Android, for iOS returns null.
+  final double? confidence;
+
+  // The angle of the rotation of the recognized symbol.
+  // Only available in Android, for iOS returns null.
+  final double? angle;
+
+  /// Constructor to create an instance of [TextSymbol].
+  TextSymbol({
+    required this.text,
+    required this.boundingBox,
+    required this.recognizedLanguages,
+    required this.cornerPoints,
+    required this.confidence,
+    required this.angle,
+  });
+
+  /// Returns an instance of [TextSymbol] from a given [json].
+  factory TextSymbol.fromJson(Map<dynamic, dynamic> json) {
+    final text = json['text'];
+    final rect = RectJson.fromJson(json['rect']);
+    final recognizedLanguages =
+        _listToRecognizedLanguages(json['recognizedLanguages']);
+    final points = _listToCornerPoints(json['points']);
+    final confidence = json['confidence'];
+    final angle = json['angle'];
+    return TextSymbol(
+      text: text,
+      boundingBox: rect,
+      recognizedLanguages: recognizedLanguages,
+      cornerPoints: points,
+      confidence: confidence,
+      angle: angle,
+    );
   }
 }