+++ packages/google_mlkit_commons/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.4.0
+
+* __BREAKING:__ Update InputImage.
+* __BREAKING:__ Rename InputImageData -> InputImageMetadata.
+* __BREAKING:__ Remove InputImagePlaneMetadata.
+* Only support nv21 for Android and bgra8888 for iOS.
+
 ## 0.3.0
 
 * Update dependencies.